@@ -8,12 +8,12 @@ import 'package:stacked/stacked.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'apiservice/api_provider.dart';
 import 'crc32.dart';
-import 'crypto/gost_hash.dart';
+import 'crypto_non_null/gost_hash.dart';
 import 'model/deep_link_response.dart';
 
 
 class LoginViewModel extends BaseViewModel{
-  BuildContext context;
+  late BuildContext context;
   var _apiProvider = ApiProvider();
   var _sendController = new TextEditingController();
   TextEditingController get sendController => _sendController;
@@ -33,9 +33,9 @@ class LoginViewModel extends BaseViewModel{
   Color _color=Colors.green;
   Color get color=>_color;
 
-  String documentId;
-  String siteId;
-  String challange;
+  String? documentId;
+  String? siteId;
+  String? challange;
 
   String responseSend="";
   String successMessage="";
@@ -97,9 +97,9 @@ class LoginViewModel extends BaseViewModel{
     _successStatus=false;
     notifyListeners();
     print("///////////\nsiteId: $siteId \n documentId: $documentId \n///////////");
-    print("///////////\nchallange: $challange \n challange hash: ${GostHash.hashGost(challange)}\n///////////");
-    var docHash = GostHash.hashGost(challange);
-    var code = siteId + documentId + docHash;
+    print("///////////\nchallange: $challange \n challange hash: ${GostHash.hashGost(challange!)}\n///////////");
+    var docHash = GostHash.hashGost(challange!);
+    var code = siteId! + documentId! + docHash;
     var crc32 = Crc32.calcHex(code);
     code += crc32;
     print("Deep Code $code");
